Initial commit of the chrome extension
Current functionality support sending an alert to the screen
with the time stamp of a youtube video when the hotkey
ctrl+shift+l has been pressed.

Also includes code to send a POST request to the ec2 server when
the hotkey is pressed. On the server side this request initiates
a call to download the video and subtitles.