Extension runs exclusively on youtube WATCH pages
Uses the older regex parser to determine if a page is some kind of
youtube page with a watch link. To pick up the URL and pass it through
the parser, we have to use the async chrome.tabs.query call. This means
that acting on this we have to put all the logic INTO the tabs query
callback function.

Close issue #5