Fixed issues with tabs and LFs chars appearing
Had to parse the JSON to make sure that newline and other unicode
is parsed correctly. For good measure also changed the datatype
response to be text/plain but not sure if it's necessary.

Also made changes to the server so it sends back octet but they
don't seem to effect what is going on here.